[API Review] QuadPoints in Highlight Annotation
This change is done for the following feedback by API council:

1. FreeTextAnnotation#setBounds. The parameter is already annotated as
NonNull. The API will generate automatic documentation. Given this,
there is no need to add the comments saying it throws a
NullPointerException if the parameter is null.

2. The "setBounds" method in StampAnnotation and FreeTextAnnotation sets
a single object, while in HighlightAnnotation it sets a List object.
Would it make sense to change and align the method names with the
parameters passed?

Flag: android.graphics.pdf.flags.enable_edit_pdf_annotations
Bug: 403591776
Test: atest CtsPdfModuleTestCases
Change-Id: If3c1db4dff02e5f1154ff5e76d00a63a43e6eede