Handle incorrect Media type assignment for hidden files and paths
In MediaProvider, we assign MEDIA_TYPE_NONE to files that are either hidden or located in hidden directories. However, MimeTypeFixHandler didn’t consider this earlier. This change adds checks to handle such cases. If the file is hidden or under a hidden path, or if it's an image used as album art, we now assign MEDIA_TYPE_NONE as expected

Test: atest MimeTypeFixHandlerTest
Test: atest PhotoPickerTest
Test: atest ModernMediaScannerText
Flag: com.android.providers.media.flags.enable_mime_type_fix_for_android_15
Bug: 376910932
Change-Id: I6fba82ae20468e1853358e725cc0935e1d26d08b