Handle the edge case where Photopicker package does not exist
In case the Photopicker package is not installed a profile, currently
if the user switchs profile Photopicker would crash because it tries to
create context for the Photopicker package which would not exist.

This change adds a fallback for such scenarios to create a system
context in order to get the ContentResolver.

Bug: 407932166
Test: atest ContextTest && atest DataServiceImplTest
Flag: EXEMPT bug fix
Change-Id: I19b152064acb231543bb6d286716bf16aad7c403