Conditonally send "Downloads" as pinned local album.
When the local media provider capabilities flag is enabled, we do not
need to display "Downloads" as part of pinned local albums. Instead it
is part of the "From this device" collection.

In addition to this, the changes also include fixes to localize display
name for merged albums, namely "Favorites" and "Downloads".

Bug: 404692577
Test: atest ExternalDbFacadeTest
Flag: com.android.providers.media.flags.enable_local_media_provider_capabilities
Change-Id: I8d0ea973dafb2f5309f866cd21d43df7b58d2504