Use the framework-mediaprovider impl lib only for the implementation.
`libs` also adds it to the dependencies of the autogenerated stub
modules, causing the stubs to depend on an implementation library.

This is necessary to avoid cyclic dependencies if/when R8 is used.

This also fixes the SDK generation to not include two @hide annotations
in the API.

Test: m nothing
Bug: 418247371
Flag: EXEMPT UNSUPPORTED_LANGUAGE
Change-Id: I6216652b16f480840b931c889f2eab0a82338106