Introduce dedicated background and foreground thread pools
This commit introduces BackgroundThreadPool and ForegroundThreadPool
to standardize the execution of asynchronous tasks within the
application. It also includes ThreadFactoryUtil,
a new utility class, to manage thread creation.

The primary motivation for these dedicated pools is to:

- Improve Responsiveness - By offloading backgroud tasks to the
  background or applying timeouts to the foregroud
- Optimize Resource Management - Provide controlled concurrency and
  prevent resource exhaustion due to unmanaged threads.

Bug: 425659533
Test: atest ThreadPoolTest
Flag: EXEMPT simply adds new thread pool config but it's not being used
anywhere

Change-Id: Idc409dc7cdcad1b50f17a9394d3fdeaebdd212b9