fix: Ensure CMP remains in the foreground while openMedia is running.
Given the oneway nature of the AsyncContentProvider#openMedia, its
likely that long running operations in the CMP will result in the CMP
being backgrounded if its active refcount is zero. Thus, acquire a
client for the CMP and hold onto it until the openMedia call either
times out or completes.

Bug: b/427813183
Test: presubmits
Flag: EXEMPT bug fix
Change-Id: I94a22550dfb0e289ec2093fcfbbf1446828113fd