fix(mediagrid): Prevent pinch-to-zoom on grid headers
The pinch-to-zoom gesture was being triggered when initiated on header
elements in the media grid. This change prevents the gesture from
activating on anything other than the media items themselves.

The `pinchToZoom` modifier is updated to allow the consumer to cancel
the gesture at the `PinchToZoomEvent.Started` event by returning `true`.
This prevents any `Changed` events from being emitted.

The `MediaGrid` now uses this functionality to check if a pinch gesture
starts on a header and cancels it accordingly. A unit test is added to
verify this new early termination behavior.

Bug: b/432181188
Test: atest PhotopickerTests:PinchToZoomTest
Flag: EXEMPT bug_fix
Change-Id: I2ef0f1b71fff30eb3957557bb65b6ab9ae16bcc2