feat(preview): Add gesture to close preview screen
This change introduces a pinch-to-zoom-out gesture to the full-screen
preview. When a user performs this gesture, the application navigates
back, effectively closing the preview screen.

This behavior is controlled by the `MEDIA_GRID_TOUCH_FEATURES_ENABLED`
flag.

Adds an integration test to verify that performing a pinch gesture on
the preview correctly pops the back stack and returns to the previous
screen.

Bug: b/432183267
Test: atest PhotopickerTests:PreviewFeatureTest
Flag: EXEMPT bug fix
Change-Id: Ia6858b7417d92345b85545b1f67f3bae729f0ed9