PP Search :Update face thumbnail spacing based on count
With ag/34384225 the face suggestions thumbnail horizontal alignment was
changed to SpaceAround. This fixes the original issue of the thumbnails
not visible fully when exactly max limit items are displayed.
However if the count of face suggestions is less than max limit(6) the
spacing will be too large between the thumbnails.
With this change whenever number of face suggestions is less than max
limit then space width of 8dp is set between the items and only if the
count of the face suggestions is max limit then set to SpaceAround so
that all items have equal spacing to fit within the row width.

screenshot:
go/ss/BaDjCScWqgFP5MU.png

Bug: 416659019
Test: Manual
Flag: EXEMPT Bug fix
Change-Id: Ic0f81b35c643ee28bdc3c159874f48539725c3c1