For search bar remove fixed width for profile selector
This fixes the Misaligned PhotoPicker search bar by removing the fixed
width set for profile selector.
When search bar is enabled setting a fixed width for profile selector
reserves an empty space even when mutilple profiles are not avaialble.
This is removed so that only when a personal profile is avaialble there
is not space set for profile selector button

Bug: 433395005
Test: Manual
Flag: EXEMPT bug fix
Change-Id: I8a1f227d04e9ecfc93eba9c63ea7f00be451e70f