Add search with provider string resource.
Adds a new string for the search bar that includes the cloud provider's
name. This string is used as a placeholder text in the search bar when a
cloud provider is available.
The cloud provider with search enabled is provided by the
SearchDataService as a flow, which is passed on to the SearchViewModel
and final observed in UI.

Bug: 402404223
Test: atest SearchFeatureTest
Test: atest SearchViewModelTest
Test: atest SearchDataServiceImplTest
SearchBar-SS: http://screen/9n5sujTj5rGK4YN
Flag: EXEMPT conditional string resource with backup
Change-Id: I5e0308f8636ffd99554fbd7f0f7097b9fcfbf2e2