fix(photopicker): Handle RemoteException in RemoteSurfaceController
The RemoteSurfaceController makes binder calls to a remote service. If
the remote process dies, these calls can throw a RemoteException, which
was previously unhandled and would crash the Photopicker.

This change wraps all calls to the remote controller in a try-catch
block to gracefully handle RemoteException and logs an error if one
occurs. This also resolves the lint warnings that were suppressed in the
now-deleted baseline file.

Bug: b/408391418
Test: presubmits and m
Flag: EXEMPT, minor refactors
Change-Id: I5c4aa1769198c54fff70158d24f0adc7f71ab975