fix(ui): prevent selection bar button text from wrapping
This change ensures the button text remains on a single line by
truncating it with an ellipsis if it exceeds the available space.

Additionally, the secondary action in the selection bar is now flexible
in width, preventing it from pushing the primary "Done" button out of
view in narrow screen configurations.

Bug: b/425717559
Test: manually && presubmits
Flag: EXEMPT bug_fix
Change-Id: I50a8aab36d0d2e7f02b0ba3701e8e583ebb9d60a