feat(photopicker): Add desktop mode for Photopicker
This change introduces a desktop UI for the Photopicker, which
is activated on devices with the `FEATURE_PC` system feature and when
the `enable_photopicker_desktop` flag is on.

Instead of a bottom sheet, this new mode renders the Photopicker UI
directly on a `Surface` within the activity. A new `PhotopickerDesktop`
composable was created to handle this layout, and `MainActivity` now
conditionally chooses the appropriate UI implementation at runtime.

Bug: b/438571449
Test: presubmits, will be verified on desktop device.
Flag: com.android.providers.media.flags.enable_photopicker_desktop
Change-Id: I4e58fa1fe5fd0fa11ffcd32c3680ba40b90a4dec