Floor the selection count in picker to zero
Ensure that the count of selected items is never less than zero.
GrantsAwareSet implements Set. However, currently in edge cases
it's size can be a negative integer which is causing Photopicker
to crash.

This change ensures that the size of Photopicker cannot be lower
than zero.

Bug: 438323490
Bug: 409769926
Test: atest GrantsAwareSelectionTest
Flag: EXEMPT bug fix
Change-Id: I7efd15f738d1e97d18bdbefdfad17f8199ab15f8