Enable media sets pagination
While the photopicker UI was expecting media sets in paginated form, the backend was
not respecting the same.
This change serves the requested media sets to the photopicker UI in a
paginated form by resepcting the paging params which it was not taking
into consideration previously.

-MediaSetsQuery to serve the new query with paging params
-MediaSetaDataBaseUtil's getMediaSets() updated to serve the requested
count of media sets and send back paginated keys.
-Added corresponding tests

Bug: 411126346

Test: atest MediaSetsDataBaseUtilsTest

Flag: EXEMPT BUGFIX

Change-Id: Ice67344aedd09ff3b3111f77847a5bebcea21d06