Fix: Consolidate the semantic of profile selector.
Due to the presence of separate semantic for the profile selector icon,
the Switch Access was incorrectly focussing on the icon which does not
have any onClick action defined.

This change consolidates the semantic information onto the button itself
by moving the contentDescription and onClick to the parent
FilledTonalButton.

Bug: 439455833
Test: http://screencast/cast/NTM1MDY1MDYzNTM1NDExMnxlODM1ZDEwMS0wZg
Flag: EXEMPT BUGFIX
Change-Id: I26ec01bf7c21583de767c30108a41dcf919493d4