Fix: Announce selection state for grid items
When a user selects an item in the PhtooPicker grid, the content
description does not change to reflect the 'selected' state. This makes
it difficult for users of accessibility services to know which items are
selected.

This change introduces new string resources for the selected state and
updates the content description logic to use them when an item is
selected. The same behaviour is replecated in Preview UI.

Bug: 417920380
Test: atest MediaGridTest
Test: atest PreviewFeatureTest
Screencast: http://screencast/cast/NTM1MDY1MDYzNTM1NDExMnxlODM1ZDEwMS0wZg
Flag: EXEMPT BUGFIX
Change-Id: Ic4672347a89b34f275f9afa677327368ea5b206b