Fix select button in preview for single-select
This change introduces a MediaSelectionConfirmed event to handle the
final selection step directly from the preview screen in single-select
mode. This fixes a regression where the select button became
non-functional after a previous refactoring which added confirmation
step in single step mode.

The MainActivity now listens for this event to finalize the selection
and finish the activity.

Bug: 449079350
Test: atest PreviewFeatureTest
Flag: EXEMPT BUGFIX
Change-Id: I9a20ea461b29be82b43f439a7609a1daa555d9d7