Suspend the GridDragSelectNode coroutine when not in use
Previously GridDragSelectNode which is attached to the main grid
would run a continous loop to check if it needs to scroll the grid
for drag to select functionality, even when it is not in use.

This change makes use of snapshotFlows to suspend the coroutine when
drag to select feature is not in use to reduce unnecessary resource
usage.

Bug: 444588830
Test: atest GridDragSelectTest
Flag: EXEMPT BUGFIX
Change-Id: Iced5cace62fa9315c4b09d126c2df593519a2872