[Fix] Add selected a11y labels for navbar buttons.
This change adds the information about the selected state of the navbar
tabs to the semantics, so that it is announced when using talkback.

Tests have been added to verify this semantic state based on whether the
current tab is the selected one or not.

Bug: 438902628
Screenshot: http://go/scrcast/NDg4NzEzNjAwMzE2MjExMnxhN2IxNDI3NC04ZA
Test: atest NavigationBarFeatureTest
Flag: EXEMPT BUGFIX
Change-Id: I342c36352ae75ea830f98327e4db1725216d396a