Fix: Include local media when querying cloud media sets.
When querying for cloud media set, the effectiveLocalAuthority was set
to null, this resulted in local media part of a cloud album getting
excluded during the join with the picker db media table.
This change ensures that when a cloud media set is being queried for a
non-null cloud provider, local provider is included in the query so that
the local copy of the backed up media item is displayed in the media
set.

Bug: 457863735
Test: atest PickerDataLayerV2Test
Flag: EXEMPT BUGFIX
Change-Id: Icbf660180ffa9882849f216846171e065f52aee6