Replace call() with acquireUnstableContentProviderClient() calls
Mediaprovider's coupling with the photos app due to acquiring a stable
reference to its ContentProviderClient results in killing the
MediaProvider in case the photos process becomes unresponsive.

The solution is to make sure that all the references to the photos'
client held by MP are unstable which won't result in the system killing the
MP in case photos process dies.
resolver.query() holds unstable references internally.
We only need to replace all resolver.call() method references.

Bug: 444149946

Test: verified current tests, method replace calls

Flag: EXEMPT BUGFIX
Change-Id: I3ae135f13eac546d2630d37aa6e18245dfd456e4