Recompose mediagrid to resize corectly on device rotation
The mediagrid had incorrect number of cells in a  row when device is
rotated.
This change :
- Fixes the issue by updating the value of column count
variable in media grid to initial column count value whenever its values
changes.
- Introduces `previousInitialColumns` to store and track the
value of `initialColumns`. When a difference is detected between the
current `initialColumns` and the stored `previousInitialColumns`, it
indicates a device window resize config change.

Recording:
https://screencast.googleplex.com/cast/NTA1NzU0NDY2OTEwMjA4MHxkODAxNGRmMS1jYw

Bug: 431144137
Test: Manual
Flag: EXEMPT BUGFIX
Change-Id: I6f403e026bba2ccd5b843d9edefd3d1f76c76904