Refactor device storage metrics job scheduling at idle maintenance
Metric job was scheduled at device reboot earlier. Refactor to schedule
metrics logging job at device idle maintenance run.
This ensures that the work manager instance is available at the
time of job scheduling.

This logging was removed briefly as it was suspected to be the cause of
an ANR crash

Bug: 463519386
Test: EXEMPT BUGFIX
Flag: EXEMPT BUGFIX
Change-Id: If3f0ac13893819e8489741cfce9c8added513432