Recalculate window size on device rotation to fix grid resize issue in Embedded photopicker
This change fixes the bug that prevented picker's grid from resizing in
embedded mode upon device rotation.
The fix keys remember to both configuration and embedded recompose
toggle that is set on device resize. This ensures that WindowSizeClass
is re-evaluted and the mediagrid is redrawn with correct number of cells
in a row when device orientation changes in embedded scenario.

Bug: 431144137
Test: Manual
Flag: EXEMPT BUGFIX
Change-Id: I6131966b4412e761314c988faa48aae3fd61b657