Revert "Fix black background in transparent image in Full view."
This reverts commit 6e051ba62a86edd3e5526999e00affa865315e57.

Reason for revert:  The changes here set the entire preview image slot to black or white based on the device theme. Because of this, normal images that do not fill the entire slot (e.g. tall images with left-right padding or wide images with top-bottom padding) now show the background color in the empty space. This leads to the poor visual experience described in b/469015119

BUG: 469015119, 375338259
Change-Id: I5a5815171cd68aabe2bd1c5cbeb0cf4e3c5dccec